fix(log): cast to correct type

+++ src/core/cartridge.cpp
@@ -63,7 +63,7 @@ Cartridge::Cartridge(const char* filename)
         CRC32 = crc32(buf, len, CRC32);
     }
     CRC32 ^= ~0U;
-    LOGF("CRC32: %08X\n", CRC32);
+    LOGF("CRC32: %08X\n", (unsigned int)CRC32);
 }
 
 Cartridge::~Cartridge()
